Calculate grid dims based on shape of data given to streams
Depending on the shape of the data being processed by a stream (which in
turn is based on the shape of the input data), using a square shape for
the grid dimensions can potentially:
- request unnecessarily many threads
- leave unnecessarily many threads idle

This is due to the mapping of threads to pixel coordinates (i, j) in the
kernel function, and then the subsequent exclusion of certain threads
based on the assigned (i, j) coordinates.

This change still assumes (16, 16) thread-blocks, but tries to find a
grid of such thread-blocks that isn't necessarily square, so then less
threads are left idle.

This change hasn't improved the residual between the CPU and GPU
filtered results, but it has made the GPU filtering quicker to complete.