Copy an extra image row for all but the last stream
The process of median filtering requires access to neighbouring pixels
to calculate the median.

When using streams to split the data being processed, what could happen
is that:
- a pixel is being processed that exists on the bottom edge of the data
  for, say, stream A
- but due to the asynchronous copying of data from CPU to GPU via the
  streams, the next row of pixels (which is part of the data copied over
  to the GPU in a different stream, call it stream B) may not yet exist
  in GPU memory when stream A is processing its pixels

Therefore, copying over an extra row of the image for each stream (apart
from the last) will ensure that any neighbouring pixels which are on the
other side of a "stream boundary" will be in GPU memory when the stream
begins its processing.