Copy integer multiples of image rows in streams
After closer inspection (by manually reducing the number of thread-block
rows needed by 1 and checking the missing pixels in the result), it was
discovered that the number of bytes being copied/processed in a stream
didn't correspond to integer multiples of rows of the image. Ie, the
number of bytes being copied/processed by a stream was:
- some integer multiple of rows
- plus, a subset of pixels in the next row in the image

This happenened due to the integer division of the number of bytes per
stream by the number of pixels in a row in the image unintentionally
truncating the result (introduced in 72b6e44). This then seemingly
implied that an integer multiple of rows was exactly equal to the number
of bytes that a stream should process, thus implying that the number of
bytes being copied over per stream was exactly an integer multiple of
image rows.

However, there is often a remainder from this division (which was lost
in the calculation), so this is often incorrect. Therefore, the number
of bytes each stream was copying over was often more than the number of
bytes in an integer multiple of image rows. Therefore, this mistakenly
missed remainder corresponds to the extra pixels being copied over in
the streams.

Practically speaking, this mistake led to rows of thread-blocks not
mapping nicely onto rows of the image, and instead having some other
unintuitive mapping.

This change makes the mapping be the expected/more intuive mapping of
thread-blocks to rows of the image (per stream). Note that this change
hasn't decreased the residual between CPU and GPU processing, but has
slightly increased it instead (by ~100).

However, this seems like a reasonably small price to pay for having
thread-block mapping to image rows that is much more intuitive, so
that's the reasoning behind keeping ahold of this change for now.