Add stream offset to `j` pixel coord calculation
The approach that has been used to split up the image data so then it
can be processed by different streams was to process horizontal blocks
of the image. Then, each kernel function call would be given a different
horizontal block of data and a different stream to process that block of
data on.

The arrangement of the thread-blocks in the grid that is passed to the
kernel function was thus motivated by the size of each horizontal block
of the image that a stream would process.

Due to how the `i` and `j` values are calculated in the kernel function
in combination with each stream being given the same grid shape of
thread-blocks, this in particular meant that the calculation of the
`(i, j)` pixel coordinate could come out the same across different
streams. This resulted in both
- some pixels in the image not being processed at all
- some pixels in the image being processed more than once

To give more detail on the way that this happened, consider the
following:
- in stream 1, the kernel function has been given a grid of
  thread-blocks
- consider a thread in this grid with some particular set of values for
  the x and y components of `blockIdx` and `threadIdx`
- these will map onto some `i` and `j` via a calculation in the kernel
  function
- now consider stream 2; the kernel function has also been given a grid
  of thread-blocks to do processing on this stream
- in particular, the dimensions of the grid of thread-blocks being used
  for processing on stream 1 and stream 2 are the same (so then they can
  fully process the appropriately sized horizontal block of the image)
- therefore, if a thread is considered in the grid given to the kernel
  function for stream 2 to process, and it has the same position in that
  grid as another thread in the grid given to the kernel function for
  stream 1 to process, those two threads used in different streams
  actually map to the same pixel coordinate `(i, j)`
- this is because the `i` and `j` values are solely based on the x and y
  components of `blockIdx` and `threadIdx`, even across different
  streams

Therefore, this change introduces an offset to the `j` value which is
based on the offset for the stream that is currently performing
processing. Practically speaking, what this does is make sure that a
given thread within a stream calculates its `j` value so that it
processes the intended pixel inside the horizontal block of data
associated with the stream.

Another way to explain it would be that this change is ensuring that the
mapping:
- from x and y components of `blockIdx` and `threadIdx`
- to `i` and `j` pixel coordinates
is unique/one-to-one, by adding information about the stream offset to
the mapping.