Copy extra vertical slices for streams in 3D filtering
These changes are so that, when pixels on a boundary of a vertical block
in a stream are being processed, their full neighbourhood is available
to correctly perform the median calculation.

For the first stream, copy over one extra vertical slice after the
vertical block being processed in the stream.

For the middle streams, copy over two extra vertical slices:
- one *before* the vertical block being processed in the stream
- one *after* the vertical block being processed in the stream

For the last stream, copy over one extra vertical slice before the
vertical block being processed in the stream.