Benchmark w/o streaming, and two approaches w/ streaming
To start testing if the use of streaming provides a performance increase
as expected or not, several cases similar to the cases in this example
code from Nvidia
https://github.com/NVIDIA-developer-blog/code-samples/blob/7d974c5bc761c650d13b81d8c2fb311899f8564e/series/cuda-cpp/overlap-data-transfers/async.cu#L95
have been clocked using CUDA events.

The results are that both approaches which use four streams actually
peform worse than when using a single stream (the default one), so
profiling of the code needs to be done to understand why this is
happening.