Use pinned memory on host
Profiling the CUDA code revealed at least one reason for the lack of
performance increase when using four streams over one stream: the
processing in each stream was running serially. Ie, one stream was:
- copying from host to device
- executing the kernel
- copying from device to host
before another stream would begin, and there was no overlapping
whatsoever between data transfers and kernel executions.

This article
https://developer.nvidia.com/blog/how-overlap-data-transfers-cuda-cc/
pointed out that the data in host memory that is invovled in the data
transfers needs to be "pinned memory", as opposed to "paged memory".

Using pinned memory on the host side and then profiling revealed that
this then allowed some concurrency across the streams. Specifically, the
host-to-device and device-to-host data transfers were now able to
overlap with kernel executions. This in turn improved the performance of
using multiple streams compared to the previous commit.

However, it must be noted that even though using pinned memory in
conjunction with four streams is now faster than using paged memory with
four streams, it is still not faster than using a single stream when
using the test volume data.

One possible explanation for this is that the test volume data being
used is very small (~132MB when uncompressed), and so the overhead of
using multiple streams is outweighing the benefit of using multiple
streams. Therefore, testing with larger volume data will be explored in
the near future.