Use a data type to handle very large 3D data
When the dimensions of the 3D data get large enough, values like
- the product of the dimensions
- the number of bytes needing to be allocated for the 3D data
are not able to be stored in an `int`. This causes erroneous values to
be passed to CUDA memory allocation functions, and thus causing CUDA
memory errors.

To handle cases of larger 3D data in an attempt to make the 3D median
filter more generic, the use of `int` has been replaced with the use of
`unsigned long long` in the appropriate places.