Change order of data transfer and kernel execution operations to allow overlapping across CUDA streams
For an unknown reason, the data transfer and kernel execution operations
for the 3D median filter were no longer overlapping. It appears that the
order of these operations now makes a difference between overlapping and
not overlapping (whereas in the past it didn't seem to matter), though
the reason for this change is still unknown as well.

For now, the order of the operations has been changed to enable the
overlapping to occur (also unknown why this particular ordering allows
overlapping), and thus to once again make use of multiple CUDA streams.